+++ tests/src/Kernel/ResumeChainKernelTest.php
@@ -130,6 +130,7 @@ class ResumeChainKernelTest extends KernelTestBase {
       $this->container->get('scolta.index_locator'),
       $this->container->get('scolta.index_build_runner'),
       $this->container->get('queue'),
+      $this->container->get('entity_type.manager'),
     ) extends ScoltaCommands {
 
       /**